requirements.txt
chromadb==0.6.3
faiss-cpu==1.10.0
hydra-core==1.3.2
ollama==0.4.7
omegaconf==2.3.0
PyMuPDF==1.25.4
redis==5.2.1
sentence-transformers
subprocess